DNA #494: Fix for wrapping long URLs in Webkit
BE AWARE: This breaks the previous wrapping label alignment fix
(a86973d5c) which has already been merged to dev!

Committing this here for the benefit of the UX team so we all have
access to my previous work in hope that it helps us squash this bug
for good. The key to this fix is adding display:table to the the <dd>,
but this in turn kills our nicely aligned labels, which were acheived
using a slight twist (clearing only left floats) to the now standard
clearfix approach.

The CSS is full of comments as well as commented out rules. You can
also read more about the wrapping long strings of text here:

* http://css-tricks.com/snippets/css/prevent-long-urls-from-breaking-out-of-container

* http://blog.kenneth.io/blog/2012/03/04/word-wrapping-hypernation-using-css